refactor RoclRL to rockrl
