+++ .gitignore
@@ -3,3 +3,5 @@ __pycache__/
 venv
 runs/*
 *.egg-info
+build
+dist
